refactor: export tick-methods
